+++ js/get-random-number.js
@@ -1,0 +1,28 @@
+export const getRandomPositiveInteger = (a, b) => {
+  if (a < 0 || b < 0) {
+    return NaN;
+  }
+
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+export function getRandomPositiveFloat (a, b, digits = 1) {
+  if (a < 0 || b < 0 || digits < 0) {
+    return NaN;
+  }
+
+  const lower = Math.min(a, b);
+  const upper = Math.max(a, b);
+
+  // Дальше используем Math.random() для получения случайного дробного числа в диапазоне [0, 1),
+  // которое домножаем на разницу между переданными числами - это будет наша случайная дельта.
+  // После нужно сложить дельту с минимальным значением, чтобы получить итоговое случайное число.
+  const result = Math.random() * (upper - lower) + lower;
+
+  return +result.toFixed(digits);
+}
+
+export const getRandomArrayElement = (elements) => elements[getRandomPositiveInteger(0, elements.length - 1)];
