+++ js/main.js
@@ -1,78 +1,6 @@
-const type = ['palace', 'flat', 'house', 'bungalow', 'hotel'];
-const time = ['12:00', '13:00', '14:00'];
-const features = ['wifi', 'dishwasher', 'parking', 'washer', 'elevator', 'conditioner'];
-const photos = [
-  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
-  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
-  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg'
-];
-
-const getRandomPositiveInteger = (a, b) => {
-  if (a < 0 || b < 0) {
-    return NaN;
-  }
-
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
-  const result = Math.random() * (upper - lower + 1) + lower;
-  return Math.floor(result);
-};
-
-function getRandomPositiveFloat (a, b, digits = 1) {
-  if (a < 0 || b < 0 || digits < 0) {
-    return NaN;
-  }
-
-  const lower = Math.min(a, b);
-  const upper = Math.max(a, b);
-
-  // Дальше используем Math.random() для получения случайного дробного числа в диапазоне [0, 1),
-  // которое домножаем на разницу между переданными числами - это будет наша случайная дельта.
-  // После нужно сложить дельту с минимальным значением, чтобы получить итоговое случайное число.
-  const result = Math.random() * (upper - lower) + lower;
-
-  return +result.toFixed(digits);
-}
-
-function getFeatures() {
-  const maxLength = features.length;
-  const lengthOfArray = getRandomPositiveInteger(1, maxLength);
-  const array = [];
-
-  while (array.length < lengthOfArray) {
-    const indexOfEl = getRandomPositiveInteger(0, maxLength - 1);
-    const el = features[indexOfEl];
-
-    if (!array.includes(el)) {
-      array.push(el);
-    }
-  }
-  return array;
-} // Функция из гугла, но я постарался максимально разобраться в ней
-
-function getPhotos() {
-  const maxLength = photos.length;
-  const lengthOfArray = getRandomPositiveInteger(1, maxLength);
-  const array = [];
-
-  while (array.length < lengthOfArray) {
-    const indexOfEl = getRandomPositiveInteger(0, maxLength - 1);
-    const el = features[indexOfEl];
-
-    if (!array.includes(el)) {
-      array.push(el);
-    }
-  }
-  return array;
-}
-
-function getAvatar (id) {
-  const numberPhoto = id.toString().padStart(2, '0');
-  const linkAvatar = `img/avatars/user${numberPhoto}.png`;
-  return linkAvatar;
-}
-
-const getRandomArrayElement = (elements) => elements[getRandomPositiveInteger(0, elements.length - 1)];
+import {getRandomPositiveInteger, getRandomPositiveFloat, getRandomArrayElement} from './get-random-number.js';
+import {getAvatar, getFeatures, getPhotos} from './get-element.js';
+import {type, time} from './util.js';
 
 const createAd = (id) => {
   const randomPrice = getRandomPositiveInteger(0, 20000);
