+++ js/util.js
@@ -1,0 +1,8 @@
+export const type = ['palace', 'flat', 'house', 'bungalow', 'hotel'];
+export const time = ['12:00', '13:00', '14:00'];
+export const features = ['wifi', 'dishwasher', 'parking', 'washer', 'elevator', 'conditioner'];
+export const photos = [
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg'
+];
